+++ pollution_v2/src/common/connector/common.py
@@ -224,6 +224,7 @@ class ODHBaseConnector(ABC, Generic[MeasureType, StationType]):
             "offset": offset
         })
 
+        logger.debug(f"Retrieving path [{path}] with query params [{query_params}]")
         raw_data = self._get_request(path, query_params)
 
         if not isinstance(raw_data, dict):
@@ -234,21 +235,27 @@ class ODHBaseConnector(ABC, Generic[MeasureType, StationType]):
         except KeyError as e:
             raise ValueError(f"Unable to build a result page from server response: {e}")
 
-    def _get_result_list(self, path: str, query_params: Optional[dict] = None) -> List[dict]:
+    def _get_result_list(self, path: str, query_params: Optional[dict] = None, max_elements: Optional[int] = None) -> List[dict]:
         """
         Request the result list from an endpoint.
 
         :param path:
         :param query_params: Any additional query param
+        :param max_elements: If set, it limits the number of elements to be retrieved
         :return: The result list
         """
         limit = self._pagination_size
+        if max_elements is not None and max_elements < limit:
+            limit = max_elements
         offset = 0
         completed = False
 
         results: List[dict] = []
 
         while not completed:
+            if max_elements is not None and len(results) >= max_elements:
+                break
+
             if limit == -1:
                 logger.debug("Retrieving all elements")
             else:
@@ -302,32 +309,37 @@ class ODHBaseConnector(ABC, Generic[MeasureType, StationType]):
         return [self.build_station(raw_station) for raw_station in raw_stations]
 
     def get_latest_measures(self, station: Optional[Station or str] = None,
-                            period_to_include: int = None) -> List[MeasureType]:
+                            period_to_include: int = None, data_types: list[str] = None) -> List[MeasureType]:
         """
         Retrieve the last measure for the connector DataType.
 
         :param station: If set, it is possible to retrieve only the measures for the given station. It can be a string
                         representing the station code or a Station object.
         :param period_to_include: If set, it allows filtering including period; otherwise, no filter on period
+        :param data_types: If set, uses this custom list of data types to retrieve data
         :return: The list of measures
         """
 
-        where_conds = self.__build_where_conds(station, period_to_include)
+        where_conds = self._build_where_conds(station, period_to_include)
         query_params = {}
         if len(where_conds) > 0:
             query_params["where"] = f'and({",".join(where_conds)})'
 
-        logger.debug(f"Retrieving latest measures on [{type(self).__name__}] with where [{query_params['where']}]")
+        logger.debug(f"Retrieving latest measures on [{type(self).__name__}] for {self._station_type} with where [{query_params['where']}]")
 
+        req_data_types = data_types if data_types else self._measure_types
+        if data_types:
+            logger.info(f"Retrieving latest measures for data types [{data_types}]")
         raw_measures = self._get_result_list(
-            path=f"/v2/flat,node/{self._station_type}/{','.join(self._measure_types)}/latest",
+            path=f"/v2/flat,node/{self._station_type}/{','.join(req_data_types)}/latest",
             query_params=query_params
         )
 
         return [self.build_measure(raw_measure) for raw_measure in raw_measures]
 
     def get_measures(self, from_date: datetime, to_date: datetime, station: Optional[Station or str] = None,
-                     period_to_include: int = None, conditions: list[str] = None, measure_types: list[str] = None) -> List[MeasureType]:
+                     period_to_include: int = None, conditions: list[str] = None, measure_types: list[str] = None,
+                     limit: Optional[int] = None) -> List[MeasureType]:
         """
         Retrieve the measures for the connector DataType in the given interval.
 
@@ -338,6 +350,7 @@ class ODHBaseConnector(ABC, Generic[MeasureType, StationType]):
         :param period_to_include: If set, it allows filtering including period; otherwise, no filter on period
         :param conditions: Additional conditions to be added to the where clause
         :param measure_types: If set, it passed the measure types to retrieve data
+        :param limit: If set, it limits the number of measures to be retrieved
         :return: the list of Measures
         """
 
@@ -347,22 +360,22 @@ class ODHBaseConnector(ABC, Generic[MeasureType, StationType]):
         iso_from_date = from_date.isoformat(timespec="milliseconds")
         iso_to_date = to_date.isoformat(timespec="milliseconds")
 
-        where_conds = self.__build_where_conds(station, period_to_include, conditions)
+        where_conds = self._build_where_conds(station, period_to_include, conditions)
         query_params = {}
         if len(where_conds) > 0:
             query_params["where"] = f'and({",".join(where_conds)})'
 
         logger.debug(f"Retrieving measures on [{type(self).__name__}] from date [{iso_from_date}] "
-                     f"to date [{iso_to_date}] with where [{query_params['where']}]")
+                     f"to date [{iso_to_date}] for {self._station_type} with where [{query_params['where']}]")
         measure_types = measure_types if measure_types else self._measure_types
         raw_measures = self._get_result_list(
             path=f"/v2/flat,node/{self._station_type}/{','.join(measure_types)}/{iso_from_date}/{iso_to_date}",
-            query_params=query_params
+            query_params=query_params, max_elements=limit
         )
 
         return [self.build_measure(raw_measure) for raw_measure in raw_measures]
 
-    def __build_where_conds(self, station: Optional[Station or str] = None,
+    def _build_where_conds(self, station: Optional[Station or str] = None,
                             period_to_include: int = None, conditions: list[str] = None) -> List[str]:
 
         where_condition = []