+++ pollution_v2/src/dags/aiaas_pollution_dispersal.py
@@ -8,16 +8,20 @@ from datetime import timedelta, datetime
 
 from airflow.decorators import task
 from airflow.utils.trigger_rule import TriggerRule
+from redis.client import Redis
 
-from dags.common import StationsDAG
+from common.cache.computation_checkpoint import ComputationCheckpointCache
+from dags.common import TrafficStationsDAG
 from common.connector.collector import ConnectorCollector
 from common.data_model.common import Provenance
-from common.data_model import Station, TrafficSensorStation
+from common.data_model import TrafficSensorStation
 from common.settings import (ODH_MINIMUM_STARTING_DATE, PROVENANCE_ID, PROVENANCE_LINEAGE,
                              PROVENANCE_NAME_POLL_ELABORATION, PROVENANCE_VERSION, AIRFLOW_NUM_RETRIES,
                              DAG_POLLUTION_DISPERSAL_EXECUTION_CRONTAB, POLLUTION_DISPERSAL_STATION_MAPPING_ENDPOINT,
                              DAG_POLLUTION_DISPERSAL_TRIGGER_DAG_HOURS_SPAN,
-                             ODH_COMPUTATION_BATCH_SIZE_POLL_DISPERSAL)
+                             ODH_COMPUTATION_BATCH_SIZE_POLL_DISPERSAL, COMPUTATION_CHECKPOINT_REDIS_HOST,
+                             COMPUTATION_CHECKPOINT_REDIS_PORT, COMPUTATION_CHECKPOINT_REDIS_DB,
+                             POLLUTION_DISPERSAL_STARTING_DATE)
 from pollution_dispersal.manager.pollution_dispersal import PollutionDispersalManager
 
 # see https://airflow.apache.org/docs/apache-airflow/stable/authoring-and-scheduling/dynamic-task-mapping.html
@@ -37,7 +41,7 @@ default_args = {
 
 THIS_DAG_ID = "pollution_dispersal"
 
-with StationsDAG(
+with TrafficStationsDAG(
     THIS_DAG_ID,
 
     # execution interval if no backfill step length on date increment if backfill (interval determined by first slot
@@ -69,14 +73,13 @@ with StationsDAG(
         """
 
         checkpoint_cache = None
-        # TODO: decide if we want to use the checkpoint cache
-        # if COMPUTATION_CHECKPOINT_REDIS_HOST:
-        #     logger.info("Enabled checkpoint cache")
-        #     checkpoint_cache = ComputationCheckpointCache(Redis(host=COMPUTATION_CHECKPOINT_REDIS_HOST,
-        #                                                         port=COMPUTATION_CHECKPOINT_REDIS_PORT,
-        #                                                         db=COMPUTATION_CHECKPOINT_REDIS_DB))
-        # else:
-        #     logger.info("Checkpoint cache disabled")
+        if COMPUTATION_CHECKPOINT_REDIS_HOST:
+            logger.info("Enabled checkpoint cache")
+            checkpoint_cache = ComputationCheckpointCache(Redis(host=COMPUTATION_CHECKPOINT_REDIS_HOST,
+                                                                port=COMPUTATION_CHECKPOINT_REDIS_PORT,
+                                                                db=COMPUTATION_CHECKPOINT_REDIS_DB))
+        else:
+            logger.info("Checkpoint cache disabled")
 
         connector_collector = ConnectorCollector.build_from_env()
         provenance = Provenance(PROVENANCE_ID, PROVENANCE_LINEAGE, PROVENANCE_NAME_POLL_ELABORATION, PROVENANCE_VERSION)
@@ -111,10 +114,6 @@ with StationsDAG(
                 logger.warning(f"Duplicate station mapping entries for traffic_station_id [{station_id}]: "
                                f"First weather_station_id [{previous}], Second weather_station_id [{current}]")
             station_mapping[x["traffic_station_id"]] = x["weather_station_id"]
-
-        # Serialization and deserialization is dependent on speed.
-        # Use built-in functions like dict as much as you can and stay away
-        # from using classes and other complex structures.
         station_dicts = []
         unique_station_codes = set()
         for station in stations_list:
@@ -134,12 +133,15 @@ with StationsDAG(
             station_dicts.append(station.to_json())
 
         logger.info(f"Retrieved {len(station_dicts)} stations")
-        if len(unique_station_codes) == len(station_mapping):
+        difference = unique_station_codes - set(station_mapping.keys())
+        if len(difference) > 0:
             logger.info("All stations mapped in enabled domains have been retrieved: " + str(unique_station_codes))
         else:
-            difference_str = str(unique_station_codes - set(station_mapping.keys()))
-            logger.warning("Some stations mapped in enabled domains have not been retrieved: " + difference_str)
+            logger.warning("Some stations mapped in enabled domains have not been retrieved: " + str(difference))
 
+        # Serialization and deserialization is dependent on speed.
+        # Use built-in functions like dict as much as you can and stay away
+        # from using classes and other complex structures.
         return station_dicts
 
 
@@ -156,14 +158,42 @@ with StationsDAG(
 
         manager = _init_manager()
 
+        min_from_date, max_to_date = dag.init_date_range(POLLUTION_DISPERSAL_STARTING_DATE, None)
+
         computation_start_dt = datetime.now()
         logger.info(f"Running computation")
-        manager.run_computation_for_multiple_stations(stations)
+        manager.run_computation(stations, min_from_date, max_to_date, ODH_COMPUTATION_BATCH_SIZE_POLL_DISPERSAL,
+                                keep_looking_for_input_data=True)
 
         computation_end_dt = datetime.now()
         logger.info(f"Completed computation in [{(computation_end_dt - computation_start_dt).seconds}]")
 
 
+    @task(trigger_rule=TriggerRule.ALL_DONE)
+    def whats_next(already_processed_stations, **kwargs):
+        """
+        Checks if there are still data to be processed before ending DAG runs
+
+        :param already_processed_stations: the stations already processed (not used)
+        """
+        manager = _init_manager()
+
+        def has_remaining_data(starting_date: datetime, ending_date: datetime) -> bool:
+            """
+            Determines if there are still enough data to be processed for another DAG run on the specific station.
+
+            :param starting_date: the date on which data availability starts
+            :param ending_date: the date on which data availability ends
+            :return: true if there are enough data to run another DAG on this station
+            """
+            return (ending_date - starting_date).total_seconds() / 3600 > DAG_POLLUTION_DISPERSAL_TRIGGER_DAG_HOURS_SPAN
+
+        min_from_date, _ = dag.init_date_range(POLLUTION_DISPERSAL_STARTING_DATE, None)
+        dag.trigger_next_dag_run(manager, dag, has_remaining_data, ODH_COMPUTATION_BATCH_SIZE_POLL_DISPERSAL,
+                                 True, True, True, min_from_date=min_from_date, **kwargs)
+
     tmp = get_stations_list()
 
     processed_stations = process_stations(tmp)
+
+    whats_next(processed_stations)